Bitmap calculation moved into the hooked onCreate() method
After the snapchat update, it is much more sensible to do the calculation in the hooked method, because now we can pass a Bitmap Object to Snapchat directly, instead of putting a compressed version of the image into a byte array.
We also introduce a check if the current Bitmap was already created, because onCreate() is also called when the screen is rotated. But the Bitmap shouldn't be recalculated every time the phone is rotated...