Fix to work with newest Snapchat 3.0.7 as of 2013-08-22. Proper video deletion.
Quite a few things have changed in Snapchat, so I had to modify the way the SnapPreviewFragment is started. Calling onSnapCaputerEvent with our own SnapCapturedEvent that holds the media doesn't work any more, because then a un-visibiliy call to the camera is made, which crashed Snapchat, as the camera is not loaded yet. Instead, the SnapPreviewFragment is put in the appropriate member in the LandingPageActivity object and then the preview fragment is created manually in the initSnapPreviewFragment() method. The preview fragment then pulls the Snap from the LandingPageActivity automatically.

Also, video deletion prevention is fixed. Now every request to delete a video file is directly checked against a pattern that indicates whether a video is in Snapchat's own video cache, which would mean it is a just recorded video by the camera, and then deletion will be allowed. If it doesn't match the pattern, the video must be one shared with Snapshare and video deletion is prevented.